Refactor code and project structure
