+++ src/sentry_settings.cpp
@@ -1,13 +1,13 @@
-#include "settings.h"
+#include "sentry_settings.h"
 
 #include <godot_cpp/classes/config_file.hpp>
 #include <godot_cpp/classes/project_settings.hpp>
 
 using namespace godot;
 
-Settings *Settings::singleton = nullptr;
+SentrySettings *SentrySettings::singleton = nullptr;
 
-void Settings::_load_config() {
+void SentrySettings::_load_config() {
 	// Read project config since ProjectSettings singleton may not be initialized at this point.
 	Ref<ConfigFile> conf = memnew(ConfigFile);
 	Error err = conf->load("res://project.godot");
@@ -22,26 +22,26 @@ void Settings::_load_config() {
 	sample_rate = (double)conf->get_value("sentry", "config/sample_rate", 1.0);
 }
 
-void Settings::_define_setting(const String &p_setting, const Variant &p_default) {
+void SentrySettings::_define_setting(const String &p_setting, const Variant &p_default) {
 	if (!ProjectSettings::get_singleton()->has_setting(p_setting)) {
 		ProjectSettings::get_singleton()->set(p_setting, p_default);
 	}
 	ProjectSettings::get_singleton()->set_initial_value(p_setting, p_default);
 }
 
-void Settings::_define_setting(const godot::PropertyInfo &p_info, const godot::Variant &p_default) {
+void SentrySettings::_define_setting(const godot::PropertyInfo &p_info, const godot::Variant &p_default) {
 	_define_setting(p_info.name, p_default);
 	Dictionary info = (Dictionary)p_info;
 	ProjectSettings::get_singleton()->add_property_info(info);
 }
 
-void Settings::_define_project_settings() {
+void SentrySettings::_define_project_settings() {
 	_define_setting("sentry/config/dsn", String(""));
 	_define_setting("sentry/config/debug_printing", false);
 	_define_setting(PropertyInfo(Variant::FLOAT, "sentry/config/sample_rate", PROPERTY_HINT_RANGE, "0.0,1.0"), 1.0);
 }
 
-Settings::Settings() {
+SentrySettings::SentrySettings() {
 	singleton = this;
 
 	if (ProjectSettings::get_singleton()) {
@@ -51,6 +51,6 @@ Settings::Settings() {
 	_load_config();
 }
 
-Settings::~Settings() {
+SentrySettings::~SentrySettings() {
 	singleton = nullptr;
 }
