+++ src/sentry_singleton.h
