+++ src/sentry_util.cpp
