+++ src/sentry_util.h
